# openapi.json
{
  "openapi": "3.1.0",
  "paths": {
    "/api/v1/chat": {
      "post": {
        "description": "Full JSON response. Cache, smart rate limit, ResilientExecutor, optional fallback (effectiveModelAlias).",
        "operationId": "ChatController_chat",
        "parameters": [
          {
            "name": "x-request-id",
            "in": "header",
            "description": "Request correlation ID — echo request header `x-request-id` (if not empty) or `req_<uuid>`. Set by `RequestIdMiddleware` (`src/common/middleware/request-id.middleware.ts`) on all routes with middleware. The same ID as the `requestId` field in JSON (success, error, SSE `meta`).",
            "required": false,
            "schema": {
              "type": "string",
              "example": "req_123e4567-e89b-12d3-a456-426614174000"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChatRequestDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChatResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Walidacja, MODEL_ALIAS_NOT_FOUND, MODEL_NOT_ALLOWED, STREAMING_NOT_SUPPORTED (stream)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeDto"
                }
              }
            }
          },
          "401": {
            "description": "GATEWAY_KEY_MISSING",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeDto"
                }
              }
            }
          },
          "403": {
            "description": "GATEWAY_KEY_INVALID",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeDto"
                }
              }
            }
          },
          "429": {
            "description": "RATE_LIMITED lub PROVIDER_RATE_LIMITED",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeDto"
                }
              }
            }
          },
          "500": {
            "description": "INTERNAL_SERVER_ERROR, GATEWAY_KEY_NOT_CONFIGURED",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeDto"
                }
              }
            }
          },
          "502": {
            "description": "PROVIDER_UNSUPPORTED, PROVIDER_UNAVAILABLE",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeDto"
                }
              }
            }
          },
          "504": {
            "description": "PROVIDER_TIMEOUT",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "GatewayKeyAuth": []
          }
        ],
        "summary": "Standard chat",
        "tags": [
          "Chat"
        ]
      }
    },
    "/api/v1/chat/stream": {
      "post": {
        "description": "SSE (`text/event-stream`): event sequence `meta` → `delta`* → `done` (empty `{}`).\n\nSame body as POST /chat. JSON errors (ErrorEnvelope) only **before** flushHeaders.\n\nAfter stream starts, errors may be partial SSE.",
        "operationId": "ChatStreamController_streamChat",
        "parameters": [
          {
            "name": "x-request-id",
            "in": "header",
            "description": "Request correlation ID — echo request header `x-request-id` (if not empty) or `req_<uuid>`. Set by `RequestIdMiddleware` (`src/common/middleware/request-id.middleware.ts`) on all routes with middleware. The same ID as the `requestId` field in JSON (success, error, SSE `meta`).",
            "required": false,
            "schema": {
              "type": "string",
              "example": "req_123e4567-e89b-12d3-a456-426614174000"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChatRequestDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "SSE: event meta (SseMetaPayload), delta* (SseDeltaPayload), done (SseDonePayload).",
            "content": {
              "text/event-stream": {
                "schema": {
                  "type": "string",
                  "description": "Format: `event: <name>\\ndata: <json>\\n\\n`"
                },
                "examples": {
                  "meta": {
                    "value": "event: meta\ndata: {\"id\":\"gw_...\",\"provider\":\"anthropic\",\"model\":\"chat-default\",\"requestId\":\"req_...\",\"conversationId\":\"conv_...\"}"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Walidacja, MODEL_ALIAS_NOT_FOUND, MODEL_NOT_ALLOWED, STREAMING_NOT_SUPPORTED (stream)",
            "content": {
              "text/event-stream": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeDto"
                }
              }
            }
          },
          "401": {
            "description": "GATEWAY_KEY_MISSING",
            "content": {
              "text/event-stream": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeDto"
                }
              }
            }
          },
          "403": {
            "description": "GATEWAY_KEY_INVALID",
            "content": {
              "text/event-stream": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeDto"
                }
              }
            }
          },
          "429": {
            "description": "RATE_LIMITED lub PROVIDER_RATE_LIMITED",
            "content": {
              "text/event-stream": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeDto"
                }
              }
            }
          },
          "500": {
            "description": "INTERNAL_SERVER_ERROR, GATEWAY_KEY_NOT_CONFIGURED",
            "content": {
              "text/event-stream": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeDto"
                }
              }
            }
          },
          "502": {
            "description": "PROVIDER_UNSUPPORTED, PROVIDER_UNAVAILABLE",
            "content": {
              "text/event-stream": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeDto"
                }
              }
            }
          },
          "504": {
            "description": "PROVIDER_TIMEOUT",
            "content": {
              "text/event-stream": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "GatewayKeyAuth": []
          }
        ],
        "summary": "Streaming chat",
        "tags": [
          "Chat"
        ]
      }
    },
    "/api/v1/health": {
      "get": {
        "description": "LightWeight process probe. Does not require X-Gateway-Key.",
        "operationId": "HealthController_getLiveness",
        "parameters": [
          {
            "name": "x-request-id",
            "in": "header",
            "description": "Request correlation ID — echo request header `x-request-id` (if not empty) or `req_<uuid>`. Set by `RequestIdMiddleware` (`src/common/middleware/request-id.middleware.ts`) on all routes with middleware. The same ID as the `requestId` field in JSON (success, error, SSE `meta`).",
            "required": false,
            "schema": {
              "type": "string",
              "example": "req_123e4567-e89b-12d3-a456-426614174000"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HealthLivenessResponseDto"
                }
              }
            }
          }
        },
        "summary": "Liveness",
        "tags": [
          "Health"
        ]
      }
    },
    "/api/v1/health/ready": {
      "get": {
        "description": "HTTP always 200; evaluation by body.status (ready | not_ready). Does not require X-Gateway-Key.",
        "operationId": "HealthController_getReadiness",
        "parameters": [
          {
            "name": "x-request-id",
            "in": "header",
            "description": "Request correlation ID — echo request header `x-request-id` (if not empty) or `req_<uuid>`. Set by `RequestIdMiddleware` (`src/common/middleware/request-id.middleware.ts`) on all routes with middleware. The same ID as the `requestId` field in JSON (success, error, SSE `meta`).",
            "required": false,
            "schema": {
              "type": "string",
              "example": "req_123e4567-e89b-12d3-a456-426614174000"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HealthReadinessResponseDto"
                }
              }
            }
          }
        },
        "summary": "Readiness",
        "tags": [
          "Health"
        ]
      }
    }
  },
  "info": {
    "title": "AI Provider Gateway API",
    "description": "REST API: chat JSON + streaming SSE. System prompt on server side. Details: `docs/dokumentacja_api.md`.",
    "version": "0.12.0",
    "contact": {}
  },
  "tags": [
    {
      "name": "Health",
      "description": "Liveness and readiness - without X-Gateway-Key."
    },
    {
      "name": "Chat",
      "description": "Chat completions (standard + streaming SSE)."
    }
  ],
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Localhost"
    }
  ],
  "components": {
    "securitySchemes": {
      "GatewayKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Gateway-Key",
        "description": "Client key. Required for POST /chat and POST /chat/stream."
      }
    },
    "schemas": {
      "ChatMessageDto": {
        "type": "object",
        "properties": {
          "role": {
            "type": "string",
            "enum": [
              "user",
              "assistant"
            ],
            "description": "Role of the message. Must be either \"user\" or \"assistant\".",
            "example": "user"
          },
          "content": {
            "type": "string",
            "maxLength": 3000,
            "description": "Content of the message. Must be a string with a maximum length of 3000 characters.",
            "example": "Hello, how are you?"
          }
        },
        "required": [
          "role",
          "content"
        ]
      },
      "ChatParamsDto": {
        "type": "object",
        "properties": {
          "temperature": {
            "type": "number",
            "minimum": 0,
            "maximum": 2,
            "description": "Override temperature for this request. Allowed only if listed in allowOverrides for modelAlias.",
            "example": 0.7
          },
          "maxOutputTokens": {
            "type": "number",
            "minimum": 1,
            "maximum": 8192,
            "description": "Override max output tokens. Allowed only if listed in allowOverrides for modelAlias.",
            "example": 1024
          }
        }
      },
      "ChatRequestDto": {
        "type": "object",
        "properties": {
          "modelAlias": {
            "type": "string",
            "description": "Model alias to use for the request. Must be defined in the configuration.",
            "example": "claude-sonnet-4-5"
          },
          "messages": {
            "minItems": 1,
            "maxItems": 150,
            "description": "Array of messages to send in the request. Each message must have a role and content. Maximum 150 messages.",
            "example": [
              {
                "role": "user",
                "content": "Hello, how are you?"
              }
            ],
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ChatMessageDto"
            }
          },
          "params": {
            "description": "Optional generation overrides. Only fields listed in allowOverrides for this modelAlias are accepted.",
            "allOf": [
              {
                "$ref": "#/components/schemas/ChatParamsDto"
              }
            ]
          },
          "conversationId": {
            "type": "string",
            "pattern": "^conv_[0-9a-f]{8}-[0-9a-f]{4}-[1-5][0-9a-f]{3}-[89ab][0-9a-f]{3}-[0-9a-f]{12}$",
            "description": "Optional conversation ID to group multiple requests into a single conversation for metrics tracking. Generate unique ID on the client side and reuse it for all requests in the same conversation.",
            "example": "conv_123e4567-e89b-12d3-a456-426614174000"
          }
        },
        "required": [
          "modelAlias",
          "messages"
        ]
      },
      "ChatOutputTextDto": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "enum": [
              "text"
            ],
            "example": "text"
          },
          "text": {
            "type": "string",
            "description": "Odpowiedź wygenerowana przez model.",
            "example": "Odpowiedź modelu...."
          }
        },
        "required": [
          "type",
          "text"
        ]
      },
      "ChatUsageDto": {
        "type": "object",
        "properties": {
          "inputTokens": {
            "type": "number",
            "minimum": 0
          },
          "outputTokens": {
            "type": "number",
            "minimum": 0
          },
          "totalTokens": {
            "type": "number",
            "minimum": 0
          }
        }
      },
      "ChatResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "gw_01HZZZZZZZZZZZZZZZZZZZZZZ"
          },
          "provider": {
            "type": "string",
            "example": "anthropic"
          },
          "model": {
            "type": "string",
            "description": "Requested modelAlias from body",
            "example": "chat-default"
          },
          "effectiveModelAlias": {
            "type": "string",
            "description": "Only after successful fallback in YAML",
            "example": "claude-sonnet"
          },
          "output": {
            "$ref": "#/components/schemas/ChatOutputTextDto"
          },
          "usage": {
            "$ref": "#/components/schemas/ChatUsageDto"
          },
          "requestId": {
            "type": "string",
            "example": "req_01HZZZZZZZZZZZZZZZZZZZZZZ"
          },
          "conversationId": {
            "type": "string",
            "description": "Conversation ID returned to client (echo conversationId from body or conv_<uuid> when missing in request). Sentry grouping requires the same ID in body of subsequent requests — see conversation-tracking.md.",
            "example": "conv_01HZZZZZZZZZZZZZZZZZZZZZZ"
          },
          "cached": {
            "type": "boolean",
            "enum": [
              true
            ],
            "description": "Whether the response was returned from cache"
          },
          "cachedAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "id",
          "provider",
          "model",
          "output",
          "requestId",
          "conversationId"
        ]
      },
      "ErrorEnvelopeDto": {
        "type": "object",
        "properties": {
          "statusCode": {
            "type": "number",
            "example": 400,
            "minimum": 400,
            "maximum": 599
          },
          "code": {
            "type": "string",
            "example": "VALIDATION_FAILED",
            "description": "Machine-readable error code; dictionary: docs/dictionary.md"
          },
          "message": {
            "type": "string",
            "example": "modelAlias must be a string"
          },
          "requestId": {
            "type": "string",
            "example": "req_01H..."
          },
          "details": {
            "type": "array",
            "description": "Current implementation often returns []",
            "example": []
          }
        },
        "required": [
          "statusCode",
          "code",
          "message",
          "requestId"
        ]
      },
      "HealthLivenessResponseDto": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "enum": [
              "healthy"
            ],
            "example": "healthy"
          },
          "timestamp": {
            "type": "string",
            "format": "date-time",
            "example": "2026-05-19T12:00:00.000Z"
          }
        },
        "required": [
          "status",
          "timestamp"
        ]
      },
      "HealthCheckItemDto": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "enum": [
              "healthy",
              "degraded",
              "unhealthy"
            ]
          },
          "message": {
            "type": "string",
            "example": "Config is loaded"
          }
        },
        "required": [
          "status",
          "message"
        ]
      },
      "HealthReadinessChecksDto": {
        "type": "object",
        "properties": {
          "config": {
            "$ref": "#/components/schemas/HealthCheckItemDto"
          },
          "cache": {
            "description": "Response cache backend (noop/redis). Does not affect smart rate limit.",
            "allOf": [
              {
                "$ref": "#/components/schemas/HealthCheckItemDto"
              }
            ]
          }
        },
        "required": [
          "config",
          "cache"
        ]
      },
      "HealthReadinessResponseDto": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "enum": [
              "ready",
              "not_ready"
            ]
          },
          "timestamp": {
            "type": "string",
            "format": "date-time",
            "example": "2026-05-19T12:00:00.000Z"
          },
          "version": {
            "type": "string",
            "example": "1.0.0"
          },
          "uptime": {
            "type": "number",
            "example": 42,
            "minimum": 0
          },
          "checks": {
            "$ref": "#/components/schemas/HealthReadinessChecksDto"
          }
        },
        "required": [
          "status",
          "timestamp",
          "version",
          "uptime",
          "checks"
        ]
      },
      "SseMetaPayloadDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "provider": {
            "type": "string"
          },
          "model": {
            "type": "string",
            "description": "Requested modelAlias from body"
          },
          "effectiveModelAlias": {
            "type": "string",
            "description": "Model alias actually used for provider call (after fallback)"
          },
          "requestId": {
            "type": "string"
          },
          "conversationId": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "provider",
          "model",
          "requestId",
          "conversationId"
        ]
      },
      "SseDeltaPayloadDto": {
        "type": "object",
        "properties": {
          "text": {
            "type": "string"
          }
        },
        "required": [
          "text"
        ]
      }
    }
  }
}