# src/chat/dto/chat-request.dto.ts
import {
  IsString,
  IsArray,
  ValidateNested,
  ArrayMinSize,
  ArrayMaxSize,
  IsOptional,
  Matches,
} from 'class-validator';
import { ApiProperty, ApiPropertyOptional } from '@nestjs/swagger';
import { Type } from 'class-transformer';

import { ChatMessageDto } from './chat-message.dto';
import { ChatParamsDto } from './chat-params.dto';

const MAX_MESSAGES = 150;

export class ChatRequestDto {
  @ApiProperty({
    description:
      'Model alias to use for the request. Must be defined in the configuration.',
    required: true,
    example: 'claude-sonnet-4-5',
  })
  @IsString()
  modelAlias: string;

  @ApiProperty({
    type: [ChatMessageDto],
    minItems: 1,
    maxItems: MAX_MESSAGES,
    description:
      'Array of messages to send in the request. Each message must have a role and content. Maximum 150 messages.',
    required: true,
    example: [{ role: 'user', content: 'Hello, how are you?' }],
  })
  @IsArray()
  @ArrayMinSize(1)
  @ArrayMaxSize(MAX_MESSAGES)
  @ValidateNested({ each: true })
  @Type(() => ChatMessageDto)
  messages: ChatMessageDto[];

  @ApiPropertyOptional({
    description:
      'Optional generation overrides. Only fields listed in allowOverrides for this modelAlias are accepted.',
    type: ChatParamsDto,
  })
  @IsOptional()
  @ValidateNested()
  @Type(() => ChatParamsDto)
  params?: ChatParamsDto;

  @ApiPropertyOptional({
    description:
      'Optional conversation ID to group multiple requests into a single conversation for metrics tracking. Generate unique ID on the client side and reuse it for all requests in the same conversation.',
    required: false,
    example: 'conv_123e4567-e89b-12d3-a456-426614174000',
  })
  @IsOptional()
  @IsString()
  @Matches(
    /^conv_[0-9a-f]{8}-[0-9a-f]{4}-[1-5][0-9a-f]{3}-[89ab][0-9a-f]{3}-[0-9a-f]{12}$/i,
    {
      message: 'conversationId must be conv_<uuid>',
    },
  )
  conversationId?: string;
}